Apply Apache License, Version 2.0
The Apache License is a well-trusted license in the industry. It is also certified by the Open Source Institute and recognized as GPLv3 compatible by the Free Software Foundation.

It offers many protections that the CC BY-ND-NC license does not, while being much more permissive to forks. As a reference implementation for how a war system should work with towny, being permissive to forks should be encouraged.