Make an API getter static; Change CellAttackEvent timer assignment
Make the CellAttackEvent use `FlagWarAPI#getMaterialShiftTime(long)` and `setTime()` rather than assigning it using `FlagWarConfig#getFlagWaitingTime(long)`.

I suspect that this should fix the timer being per-phase, rather than the intended overall timer for an attack.

This would mean that the CellAttackEvent must be constructed every attack phase, which means KickStartCellAttackEvent is being called multiple times as well. Requires further investigation to see if it is an issue.