Fix Beacon height setting keys, add debug verbosity
So, turns out the keys in FlagWarConfig were not updated to reflect the keys in the config.yml. So, the minHeight was always equal to the top of the flag.

Honest mistake, but forever to be cemented into the halls of the git log.

Also fixed `rules.time_to_wait_after_flagged` - which was a surprise.