Fix Jitpack builds; use Java 17
Noticed this was an issue a short while back on Towny's repo as well. After doing some searching, it appears we can also designate specific jdks via SDKMAN! May as well build with the LTS release...

Thanks to @Warriorrrr for implementing the fix upstream and thus bringing this to attention.