Solves for WarZoneListener being unused.
Towny was never assigning TownBlockStatus#WARZONE in the
PlayerCacheUtil, because at some point it was made to only respect
EventWar.

This PR removes FlagWar's add/removes to TownyUniverse's #addWarZone and
#removeWarZone methods, which would trickle down into the TownyWorld
warzone map, affecting TownBlock#isWarZone.

The above TownyWorld and TownyUniverse are due to be scrapped with the
EventWar revamp anyways, and FlagWarAPI#isUnderAttack already does a
great of tracking where FlagWar flags are placed.

The WarZoneListener#isFastFailing has been rejiggered to look for the
above FlagWar#isUnderAttack and instead of testing for a
TownBlockStatus#WARZONE, it is checking the PlayerCache if an ENEMY
status is returned, which should mean it will fail fast in many
scenarios.