Add validated towny version tracker. Purge en_US-POSIX and en_GB translations. Identical to en_US for practical reasons, so no point in keeping them.
Adds Translation Keys:
- `startup.check-towny.new.pre-release`
- `startup.check-towny.new.stable-release`

If: Running Towny is newer than validated version, and minimum supported version is equal or less than the validated version; Evaluate running version for pre-release and log a warning.

If pre-release: Warn and tell users to report issues.
If release: Warn and ask user to update FlagWar.