Feature: Neighbouring plots protect from attack.
Adds 2 new config options, one which enables the feature and another
which sets how many neighbouring plots are required to protect a plot.

This feature is similar to the attack edge plots only, but further slows
attackers by requiring that they attack from a corner and chip away at a
town in order to get to their homeblock.

The feature is off by default, but when it is enabled it will require 3
neighbouring plots to grant protection.

Closes #74.